requirements.txt
alembic==1.13.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
auth==0.5.3
BareNecessities==0.2.8
blinker==1.7.0
boto3==1.34.39
botocore==1.34.39
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
eventlet==0.34.2
falcon==3.1.3
fastapi==0.109.2
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jwt==1.3.1
Mail==2.1.0
Mako==1.3.0
MarkupSafe==2.1.3
mistune==3.0.2
mongoengine==0.27.0
MouseInfo==0.1.3
numpy==1.26.3
packaging==23.2
pandas==2.1.4
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyGetWindow==0.0.9
PyJWT==2.8.0
pymongo==4.6.1
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
s3transfer==0.10.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
Werkzeug==3.0.1
WTForms==3.1.2
